Added PointWindow Class and Menu Bars
Added a layer on top of the Window class called PointWindow. PointWindow is used to handle more specific things than just window creation. It will be able to handle it's own messages and on top of that have almost all the code tailored to it's window specific needs like adding in GUI elements and handling messages.